Split CanvasViewController file into multiple ones.
