+++ LinkedIdeas/CanvasViewController+MouseEvents.swift
@@ -1,0 +1,139 @@
+//
+//  CanvasViewController+MouseEvents.swift
+//  LinkedIdeas
+//
+//  Created by Felipe Espinoza Castillo on 16/02/2017.
+//  Copyright © 2017 Felipe Espinoza Dev. All rights reserved.
+//
+
+import Cocoa
+
+// MARK: - CanvasViewController+MouseEvents
+
+extension CanvasViewController {
+  override func mouseDown(with event: NSEvent) {
+    Swift.print("\n")
+    Swift.print("[mouseDown]")
+    let point = convertToCanvasCoordinates(point: event.locationInWindow)
+
+    if event.isSingleClick() {
+      if let clickedElements = clickedElements(atPoint: point) {
+        Swift.print("[mouseDown][singleClick] clicked concepts [\(clickedConcepts)]")
+
+        if !currentState.isSimilar(to: .multipleSelectedElements(elements: [Element]())) {
+          safeTransiton {
+            try stateManager.toSelectedElement(element: clickedElements.first!)
+          }
+        }
+
+      } else {
+        Swift.print("[mouseDown][singleClick] no clicked concepts")
+
+        safeTransiton {
+          try stateManager.toCanvasWaiting()
+        }
+
+      }
+    } else if event.isDoubleClick() {
+      if let element = clickedSingleElement(atPoint: point) {
+        if !element.isEditable {
+          safeTransiton { try stateManager.toEditingElement(element: element) }
+        } else {
+          safeTransiton { try stateManager.toSelectedElement(element: element) }
+        }
+      } else {
+        safeTransiton { try stateManager.toNewConcept(atPoint: point) }
+      }
+    }
+  }
+
+  override func mouseDragged(with event: NSEvent) {
+    if dragCount <= 3 {
+      Swift.print("[mouseDragged]")
+      dragCount += 1
+    }
+
+    let point = convertToCanvasCoordinates(point: event.locationInWindow)
+
+    // Decision: which actions to trigger
+    // given the context and events that happen
+
+    switch currentState {
+    case .selectedElement(let element):
+      guard let concept = element as? Concept else {
+        return
+      }
+
+      if isDragShiftEvent(event) {
+        creationArrowForLink(toPoint: point)
+        didShiftDragStart = true
+        if let hoveredConcepts = clickedConcepts(atPoint: point) {
+          select(elements: hoveredConcepts)
+        } else {
+          unselect(elements: document.concepts.filter { $0 != concept })
+        }
+      } else {
+        drag(concept: concept, toPoint: point)
+      }
+
+    case .multipleSelectedElements(let elements):
+      guard let concepts = elements as? [Concept] else {
+        return
+      }
+      drag(concepts: concepts, toPoint: point)
+
+    case .canvasWaiting:
+      hoverConcepts(toPoint: point)
+
+    default:
+      return
+    }
+
+    reRenderCanvasView()
+  }
+
+  override func mouseUp(with event: NSEvent) {
+    Swift.print("[mouseUp] (state = \(currentState))")
+    let point = convertToCanvasCoordinates(point: event.locationInWindow)
+
+    switch currentState {
+    case .selectedElement(let element):
+      guard let concept = element as? Concept, didDragStart else {
+        return
+      }
+
+      if isDragShiftEvent(event) {
+        if let targetConcept = clickedSingleConcept(atPoint: point) {
+          Swift.print("[mouseUp][shiftClick] (targetConcept = \(targetConcept))")
+          targetConcept.isSelected = false
+
+          let link = saveLink(fromConcept: concept, toConcept: targetConcept)
+          dismissConstructionArrow()
+          safeTransiton {
+            try stateManager.toSelectedElement(element: link)
+          }
+        } else {
+          Swift.print("[mouseUp][shiftClick] no targetConcept!")
+          safeTransiton { try stateManager.toCanvasWaiting() }
+        }
+      } else {
+        Swift.print("[mouseUp][noShift] normal drag")
+        endDrag(forConcept: concept, toPoint: point)
+      }
+    case .multipleSelectedElements(let elements):
+      guard let concepts = elements as? [Concept] else {
+        return
+      }
+      endDrag(forConcepts: concepts, toPoint: point)
+
+    case .canvasWaiting:
+      selectHoveredConcepts()
+
+    default:
+      return
+    }
+
+    Swift.print("\n")
+    resetDraggingConcepts()
+  }
+}
