+++ LinkedIdeas/CanvasViewController.swift
@@ -8,138 +8,6 @@
 
 import Cocoa
 
-protocol GraphConcept {
-  var rect: NSRect { get }
-  var attributedStringValue: NSAttributedString { get }
-  var isSelected: Bool { get set }
-}
-
-protocol GraphLink {
-  var color: NSColor { get }
-  var isSelected: Bool { get set }
-
-  var stringValue: String { get }
-  var attributedStringValue: NSAttributedString { get }
-
-  var point: NSPoint { get }
-
-  var originPoint: NSPoint { get }
-  var targetPoint: NSPoint { get }
-
-  var originRect: NSRect { get }
-  var targetRect: NSRect { get }
-}
-
-extension Concept: GraphConcept {}
-
-extension Link: GraphLink {
-  var originRect: NSRect { return origin.rect }
-  var targetRect: NSRect { return target.rect }
-}
-
-struct DrawableConcept: DrawableElement {
-  let concept: GraphConcept
-
-  func draw() {
-    drawBackground()
-    drawConceptText()
-    drawSelectedRing()
-  }
-
-  func drawBackground() {
-    NSColor.white.set()
-    NSRectFill(concept.rect)
-  }
-
-  func drawConceptText() {
-    NSColor.black.set()
-    concept.attributedStringValue.draw(at: concept.rect.origin)
-  }
-
-  func drawSelectedRing() {
-    guard concept.isSelected else {
-      return
-    }
-
-    NSColor.red.set()
-    NSBezierPath(rect: concept.rect).stroke()
-  }
-}
-
-struct DrawableLink: DrawableElement {
-  let link: GraphLink
-
-  func draw() {
-    link.color.set()
-    constructArrow()?.bezierPath().fill()
-    drawSelectedRing()
-    drawLinkText()
-  }
-
-  func drawLinkText() {
-    guard link.stringValue != "" else {
-      return
-    }
-
-    // background
-    NSColor.white.set()
-    var textSize = link.attributedStringValue.size()
-    let padding: CGFloat = 8.0
-    textSize.width += padding
-    textSize.height += padding
-
-    let textRect = NSRect(center: link.point, size: textSize)
-    NSRectFill(textRect)
-
-    // text
-    let bottomLeftTextPoint = link.point.translate(
-      deltaX: -(textSize.width - padding) / 2.0,
-      deltaY: -(textSize.height - padding) / 2.0
-    )
-    let attributedStringValue = NSAttributedString(
-      attributedString: link.attributedStringValue,
-      fontColor: NSColor.gray
-    )
-    attributedStringValue.draw(at: bottomLeftTextPoint)
-  }
-
-  func drawSelectedRing() {
-    guard link.isSelected else {
-      return
-    }
-
-    NSColor.red.set()
-    constructArrow()?.bezierPath().stroke()
-  }
-
-  func constructArrow() -> Arrow? {
-    let originPoint = link.originPoint
-    let targetPoint = link.targetPoint
-
-    if let intersectionPointWithOrigin = link.originRect.firstIntersectionTo(targetPoint),
-       let intersectionPointWithTarget = link.targetRect.firstIntersectionTo(originPoint) {
-      return Arrow(point1: intersectionPointWithOrigin, point2: intersectionPointWithTarget)
-    } else {
-      return nil
-    }
-  }
-}
-
-extension NSResponder {
-  var identifierString: String {
-    return "\(type(of: self))"
-  }
-
-  func print(_ message: String) {
-    Swift.print("\(identifierString): \(message)")
-  }
-}
-
-extension NSEvent {
-  func isSingleClick() -> Bool { return clickCount == 1 }
-  func isDoubleClick() -> Bool { return clickCount == 2 }
-}
-
 class CanvasViewController: NSViewController {
   @IBOutlet weak var canvasView: CanvasView!
   @IBOutlet weak var scrollView: NSScrollView!
@@ -278,352 +146,5 @@ class CanvasViewController: NSViewController {
 
   func isDragShiftEvent(_ event: NSEvent) -> Bool {
     return event.modifierFlags.contains(.shift) || didShiftDragStart
-  }
-}
-
-// MARK: - MouseEvents
-
-extension CanvasViewController {
-  override func mouseDown(with event: NSEvent) {
-    Swift.print("\n")
-    Swift.print("[mouseDown]")
-    let point = convertToCanvasCoordinates(point: event.locationInWindow)
-
-    if event.isSingleClick() {
-      if let clickedElements = clickedElements(atPoint: point) {
-        Swift.print("[mouseDown][singleClick] clicked concepts [\(clickedConcepts)]")
-
-        if !currentState.isSimilar(to: .multipleSelectedElements(elements: [Element]())) {
-          safeTransiton {
-            try stateManager.toSelectedElement(element: clickedElements.first!)
-          }
-        }
-
-      } else {
-        Swift.print("[mouseDown][singleClick] no clicked concepts")
-
-        safeTransiton {
-          try stateManager.toCanvasWaiting()
-        }
-
-      }
-    } else if event.isDoubleClick() {
-      if let element = clickedSingleElement(atPoint: point) {
-        if !element.isEditable {
-          safeTransiton { try stateManager.toEditingElement(element: element) }
-        } else {
-          safeTransiton { try stateManager.toSelectedElement(element: element) }
-        }
-      } else {
-        safeTransiton { try stateManager.toNewConcept(atPoint: point) }
-      }
-    }
-  }
-
-  override func mouseDragged(with event: NSEvent) {
-    if dragCount <= 3 {
-      Swift.print("[mouseDragged]")
-      dragCount += 1
-    }
-
-    let point = convertToCanvasCoordinates(point: event.locationInWindow)
-
-    // Decision: which actions to trigger
-    // given the context and events that happen
-
-    switch currentState {
-    case .selectedElement(let element):
-      guard let concept = element as? Concept else {
-        return
-      }
-
-      if isDragShiftEvent(event) {
-        creationArrowForLink(toPoint: point)
-        didShiftDragStart = true
-        if let hoveredConcepts = clickedConcepts(atPoint: point) {
-          select(elements: hoveredConcepts)
-        } else {
-          unselect(elements: document.concepts.filter { $0 != concept })
-        }
-      } else {
-        drag(concept: concept, toPoint: point)
-      }
-
-    case .multipleSelectedElements(let elements):
-      guard let concepts = elements as? [Concept] else {
-        return
-      }
-      drag(concepts: concepts, toPoint: point)
-
-    case .canvasWaiting:
-      hoverConcepts(toPoint: point)
-
-    default:
-      return
-    }
-
-    reRenderCanvasView()
-  }
-
-  override func mouseUp(with event: NSEvent) {
-    Swift.print("[mouseUp] (state = \(currentState))")
-    let point = convertToCanvasCoordinates(point: event.locationInWindow)
-
-    switch currentState {
-    case .selectedElement(let element):
-      guard let concept = element as? Concept, didDragStart else {
-        return
-      }
-
-      if isDragShiftEvent(event) {
-        if let targetConcept = clickedSingleConcept(atPoint: point) {
-          Swift.print("[mouseUp][shiftClick] (targetConcept = \(targetConcept))")
-          targetConcept.isSelected = false
-
-          let link = saveLink(fromConcept: concept, toConcept: targetConcept)
-          dismissConstructionArrow()
-          safeTransiton {
-            try stateManager.toSelectedElement(element: link)
-          }
-        } else {
-          Swift.print("[mouseUp][shiftClick] no targetConcept!")
-          safeTransiton { try stateManager.toCanvasWaiting() }
-        }
-      } else {
-        Swift.print("[mouseUp][noShift] normal drag")
-        endDrag(forConcept: concept, toPoint: point)
-      }
-    case .multipleSelectedElements(let elements):
-      guard let concepts = elements as? [Concept] else {
-        return
-      }
-      endDrag(forConcepts: concepts, toPoint: point)
-
-    case .canvasWaiting:
-      selectHoveredConcepts()
-
-    default:
-      return
-    }
-
-    Swift.print("\n")
-    resetDraggingConcepts()
-  }
-}
-
-// MARK: - KeyboardEvents
-
-extension CanvasViewController {
-  override func keyDown(with event: NSEvent) {
-    let enterKeyCode: UInt16 = 36
-
-    if event.keyCode == enterKeyCode {
-      Swift.print("enter event!")
-
-      switch currentState {
-      case .selectedElement(let element):
-        safeTransiton {
-          try stateManager.toEditingElement(element: element)
-        }
-      default:
-        Swift.print("do nothing")
-      }
-    }
-  }
-}
-
-// MARK: - CanvasViewDataSource
-
-extension CanvasViewController: CanvasViewDataSource {
-  var drawableElements: [DrawableElement] {
-    var elements: [DrawableElement] = []
-
-    elements += document.concepts.map {
-      DrawableConcept(concept: $0 as GraphConcept) as DrawableElement
-    }
-
-    elements += document.links.map {
-      DrawableLink(link: $0 as GraphLink) as DrawableElement
-    }
-
-    return elements
-  }
-}
-
-// MARK: - DocumentObserver
-
-extension CanvasViewController: DocumentObserver {
-  func documentChanged(withElement element: Element) {
-    reRenderCanvasView()
-  }
-}
-
-// MARK: - StateManagerDelegate
-
-extension CanvasViewController: StateManagerDelegate {
-  func transitionSuccesfull() {
-    reRenderCanvasView()
-  }
-
-  func transitionedToNewConcept(fromState: CanvasState) {
-    guard case .newConcept(let point) = currentState else {
-      return
-    }
-
-    commonTransitionBehavior(fromState)
-
-    showTextField(atPoint: point)
-  }
-
-  func transitionedToCanvasWaiting(fromState: CanvasState) {
-    commonTransitionBehavior(fromState)
-  }
-
-  func transitionedToCanvasWaitingSavingConcept(fromState: CanvasState, point: NSPoint, text: NSAttributedString) {
-    dismissTextField()
-    let _ = saveConcept(text: text, atPoint: point)
-  }
-
-  func transitionedToSelectedElement(fromState: CanvasState) {
-    commonTransitionBehavior(fromState)
-
-    guard case .selectedElement(let element) = currentState else {
-      return
-    }
-
-    select(elements: [element])
-  }
-
-  func transitionedToMultipleSelectedElements(fromState: CanvasState) {
-    commonTransitionBehavior(fromState)
-
-    guard case .multipleSelectedElements(let elements) = currentState else {
-      return
-    }
-
-    select(elements: elements)
-  }
-
-  func transitionedToSelectedElementSavingChanges(fromState: CanvasState) {
-    guard case .selectedElement(var element) = currentState else {
-      return
-    }
-    element.attributedStringValue = textField.attributedStringValue
-    dismissTextField()
-
-    transitionedToSelectedElement(fromState: fromState)
-  }
-
-  func transitionedToEditingElement(fromState: CanvasState) {
-    commonTransitionBehavior(fromState)
-
-    guard case .editingElement(var element) = currentState else {
-      return
-    }
-
-    element.isEditable = true
-
-    showTextField(atPoint: element.point, text: element.attributedStringValue)
-  }
-
-  private func commonTransitionBehavior(_ fromState: CanvasState) {
-    switch fromState {
-    case .newConcept:
-      dismissTextField()
-    case .editingElement(var element):
-      element.isEditable = false
-      dismissTextField()
-    case .selectedElement(let element):
-      unselect(elements: [element])
-      dismissConstructionArrow()
-    case .multipleSelectedElements(let elements):
-      unselect(elements: elements)
-    default:
-      break
-    }
-  }
-}
-
-// MARK: - Transition Actions
-
-extension CanvasViewController {
-  func select(elements: [Element]) {
-    for var element in elements { element.isSelected = true }
-  }
-
-  func unselect(elements: [Element]) {
-    for var element in elements { element.isSelected = false }
-  }
-
-  func saveConcept(text: NSAttributedString, atPoint point: NSPoint) -> Bool {
-    guard text.string != "" else {
-      return false
-    }
-
-    let newConcept = Concept(attributedStringValue: text, point: point)
-
-    document.save(concept: newConcept)
-    return true
-  }
-
-  func saveLink(fromConcept: Concept, toConcept: Concept, text: NSAttributedString? = nil) -> Link {
-    var linkText = NSAttributedString(string: "")
-    if let text = text {
-      linkText = text
-    }
-
-    let newLink = Link(origin: fromConcept, target: toConcept, attributedStringValue: linkText)
-
-    document.save(link: newLink)
-
-    return newLink
-  }
-
-  func showTextField(atPoint point: NSPoint, text: NSAttributedString? = nil) {
-    textField.frame = NSRect(center: point, size: NSSize(width: 60, height: 40))
-    textField.isEditable = true
-    textField.isHidden = false
-    if let text = text { textField.attributedStringValue = text }
-    textField.becomeFirstResponder()
-  }
-
-  func dismissTextField() {
-    textField.setFrameOrigin(NSPoint.zero)
-    textField.isEditable = false
-    textField.isHidden = true
-    textField.stringValue = ""
-
-    canvasView.window?.makeFirstResponder(canvasView)
-  }
-}
-
-// MARK: - NSTextFieldDelegate
-
-extension CanvasViewController: NSTextFieldDelegate {
-  // Invoked when users press keys with predefined bindings in a cell of the specified control.
-  func control(_ control: NSControl, textView: NSTextView, doCommandBy commandSelector: Selector) -> Bool {
-    switch commandSelector {
-    case #selector(NSResponder.cancelOperation(_:)):
-      safeTransiton {
-        try stateManager.toCanvasWaiting()
-      }
-      return true
-    case #selector(NSResponder.insertNewline(_:)):
-      switch currentState {
-      case .editingElement(let element):
-        safeTransiton {
-          try stateManager.toSelectedElementSavingChanges(element: element)
-        }
-      case .newConcept:
-        safeTransiton {
-          try stateManager.toCanvasWaiting(savingConceptWithText: control.attributedStringValue)
-        }
-      default:
-        Swift.print("[textField][enterKey] unhandled event (state = \(currentState))")
-      }
-      return true
-    default:
-      return false
-    }
   }
 }
